+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/InMemorySubjectCredentialStore.kt
@@ -12,20 +12,25 @@ class InMemorySubjectCredentialStore : SubjectCredentialStore {
     private val credentials = mutableListOf<SubjectCredentialStore.StoreEntry>()
     private val attachments = mutableListOf<SubjectCredentialStore.AttachmentEntry>()
 
-    override suspend fun storeCredential(vc: VerifiableCredentialJws, vcSerialized: String) {
-        credentials += SubjectCredentialStore.StoreEntry.Vc(vcSerialized, vc)
+    override suspend fun storeCredential(
+        vc: VerifiableCredentialJws,
+        vcSerialized: String,
+        scheme: ConstantIndex.CredentialScheme
+    ) {
+        credentials += SubjectCredentialStore.StoreEntry.Vc(vcSerialized, vc, scheme)
     }
 
-    override suspend fun storeCredentialSd(
+    override suspend fun storeCredential(
         vc: VerifiableCredentialSdJwt,
         vcSerialized: String,
-        disclosures: Map<String, SelectiveDisclosureItem?>
+        disclosures: Map<String, SelectiveDisclosureItem?>,
+        scheme: ConstantIndex.CredentialScheme
     ) {
-        credentials += SubjectCredentialStore.StoreEntry.SdJwt(vcSerialized, vc, disclosures)
+        credentials += SubjectCredentialStore.StoreEntry.SdJwt(vcSerialized, vc, disclosures, scheme)
     }
 
-    override suspend fun storeCredential(issuerSigned: IssuerSigned) {
-        credentials += SubjectCredentialStore.StoreEntry.Iso(issuerSigned)
+    override suspend fun storeCredential(issuerSigned: IssuerSigned, scheme: ConstantIndex.CredentialScheme) {
+        credentials += SubjectCredentialStore.StoreEntry.Iso(issuerSigned, scheme)
     }
 
     override suspend fun storeAttachment(name: String, data: ByteArray, vcId: String) {