+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/IssuerCredentialDataProvider.kt
@@ -6,7 +6,6 @@ import at.asitplus.wallet.lib.cbor.CoseKey
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.CredentialSubject
 import at.asitplus.wallet.lib.iso.IssuerSignedItem
-import at.asitplus.wallet.lib.iso.MobileSecurityObject
 import kotlinx.datetime.Instant
 
 /**
@@ -30,7 +29,7 @@ sealed class CredentialToBeIssued {
     data class Vc(
         val subject: CredentialSubject,
         val expiration: Instant,
-        val attributeType: String,
+        val scheme: ConstantIndex.CredentialScheme,
         val attachments: List<Issuer.Attachment>? = null
     ) : CredentialToBeIssued()