+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Release 3.2.0
  - Support representing credentials in both formats, ISO and W3C VC
  - Drop `CredentialScheme.credentialDefinitionName`, is now automatically converted from `vcType`
  - Add `CredentialScheme.isoNamespace` and `CredentialScheme.isoDocType` to be used for representing custom credentials according to ISO 18013-5
+ - Remove function `Holder.storeValidatedCredentials`, class `Holder.ValidatedVerifiableCredentialJws`
+ - Add member for `CredentialScheme` to various classes like `CredentialToBeIssued.Vc`, subclasses of `IssuedCredential`, subclasses of `StoreCredentialInput` and subclasses of `StoreEntry`
+ - Add parameter for `CredentialScheme` to methods in `SubjectCredentialStore`
 
 Release 3.1.0
  - Support representing credentials in [SD-JWT](https://drafts.oauth.net/oauth-selective-disclosure-jwt/draft-ietf-oauth-selective-disclosure-jwt.html) format